+++ src/scripts/components/wrapped-select.ts
@@ -67,6 +67,13 @@ export default class WrappedSelect extends WrappedElement<HTMLSelectElement> {
 
   // eslint-disable-next-line class-methods-use-this
   _optionToChoice(option: HTMLOptionElement): ChoiceFull {
+    // option.value returns the label if there is no value attribute, which can break legacy placeholder attribute support
+    if (!option.hasAttribute('value') && option.hasAttribute('placeholder')) {
+      option.setAttribute('value', '');
+      // eslint-disable-next-line no-param-reassign
+      option.value = '';
+    }
+
     return {
       id: 0,
       groupId: 0,